links logic moved to links.util.ts
