+++ src/galaxis-footer/contact-us/contact-us.component.tsx
@@ -4,6 +4,12 @@ import SocialMediaIcon from "../../models/social-media-icon.model";
 import { Box, Link, Typography } from "@mui/material";
 import AddressModel from "../../models/address.model";
 import CustomTitle from "../title/title.component";
+import {
+  areUrlsSame,
+  getCurrentDomain,
+  isValidUrl,
+  removeDomainFromUrl,
+} from "../../utils/links.util";
 
 type Props = {
   socialMediaIcons: SocialMediaIcon[];
@@ -12,22 +18,6 @@ type Props = {
 
 const ContactUs: FC<Props> = ({ socialMediaIcons: initialIcons, address }) => {
   const [socialMediaIcons, setSocialMediaIcons] = useState<SocialMediaIcon[]>(initialIcons);
-
-  function getCurrentDomain(url: string): string {
-    const domainRegex = /^(?:https?:\/\/)?(?:dev\.|staging\.)?(.*?)\//;
-    const matches = url.match(domainRegex);
-    if (matches && matches.length >= 2) {
-      return matches[1];
-    }
-    return url;
-  }
-
-  function areUrlsSame(url1: string, url2: string): boolean {
-    if (url1 === url2) {
-      return true;
-    }
-    return false;
-  }
 
   const popularItemsMapper = useCallback(
     (currentHostName: string): SocialMediaIcon[] => {
@@ -52,21 +42,6 @@ const ContactUs: FC<Props> = ({ socialMediaIcons: initialIcons, address }) => {
     },
     [socialMediaIcons]
   );
-
-  function removeDomainFromUrl(url: string): string {
-    const pathRegex = /^(?:https?:\/\/)?[^\/]*(\/.*)/;
-    const match = url.match(pathRegex);
-
-    if (match && match.length > 1) {
-      return match[1];
-    }
-    return url;
-  }
-
-  function isValidUrl(str: string): boolean {
-    const urlRegex = /^(?:https?:\/\/)?(?:www\.)?([a-zA-Z0-9-]+\.[a-zA-Z]{2,})(?:\/\S*)?$/;
-    return urlRegex.test(str);
-  }
 
   useEffect(() => {
     const hostName = "https://dev.galaxis.xyz/";
